Make map from model

Make map from model

Usage

gdm_map(
  gdm_model,
  envlayers,
  coords,
  plot_vars = TRUE,
  scl = 1,
  display_axes = FALSE,
  quiet = FALSE
)

Arguments

gdm_model

GDM model

envlayers

SpatRaster or Raster* object (LAYER NAMES MUST CORRESPOND WITH GDM MODEL)

coords

data frame with x and y coordinates

plot_vars

whether to create PCA plot to help in variable and map interpretation (defaults to TRUE)

scl

constant for rescaling variable vectors for plotting (defaults to 1)

display_axes

display PC axes text, labels, and ticks (defaults to FALSE)

quiet

whether to operate quietly and suppress the output of tables and figures (defaults to FALSE)

Value

GDM RGB map

See also