Create gt table of RDA results

Create gt table of RDA results

Usage

rda_table(
  cor_df,
  sig = 0.05,
  sig_only = TRUE,
  top = FALSE,
  order = FALSE,
  var = NULL,
  nrow = NULL,
  digits = 2
)

Arguments

cor_df

dataframe of correlation results output from rda_cor

sig

if outlier_method = "p", the significance level to use to identify SNPs (defaults to 0.05)

sig_only

whether to only include loci with p-values less than sig (defaults to TRUE)

top

whether to only include only keep the top variable for each snp in the table by the strength of the correlation (defaults to FALSE)

order

whether to order by the magnitude of the correlation (defaults to FALSE)

var

which variables to include (defaults to including all variables)

nrow

number of rows to display (defaults to displaying all rows)

digits

number of digits to include (defaults to 2)

Value

An object of class gt_tbl

See also