Get significant outliers from RDA model

Get significant outliers from RDA model

Usage

rda_getoutliers(
  mod,
  naxes = "all",
  outlier_method = "p",
  p_adj = "fdr",
  sig = 0.05,
  z = 3,
  plot = TRUE
)

Arguments

naxes

number of RDA axes to use (defaults to "all" to use all axes), if set to "manual" a selection option with a terminal prompt will be given, otherwise can be any integer that is less than or equal to the total number of axes

outlier_method

method to determine outliers. Can either be "p" to use the p-value method from here or "z" to use the z-score based method from here

p_adj

if outlier_method = "p", method to use for p-value correction (defaults to "fdr"); other options can be found in p.adjust()

sig

if outlier_method = "p", the significance level to use to identify SNPs (defaults to 0.05)

z

if outlier_method = "z", the number of standard deviations to use to identify SNPs (defaults to 3)

plot

whether to produce scree plot of RDA axes (defaults to TRUE)

Value

results from outlier tests. If outlier_method = "p", a list of outlier SNPs, p-values, and results from rdadapt (see Capblancq et al. 2018). If outlier_method = "z", a dataframe with outlier SNP Z-scores for each axis

See also